refactor(electron-test): move mixin to util
