+++ lib/tasks/electron-test-server.js
@@ -1,6 +1,6 @@
 'use strict';
 
 const TestServerTask = require('ember-cli/lib/tasks/test-server');
-const TestMixin = require('./electron-test-mixin');
+const TestMixin = require('../utils/electron-test-mixin');
 
 module.exports = TestServerTask.extend(TestMixin);
