+++ lib/tasks/electron-test.js
@@ -1,6 +1,6 @@
 'use strict';
 
 const TestTask = require('ember-cli/lib/tasks/test');
-const TestMixin = require('./electron-test-mixin');
+const TestMixin = require('../utils/electron-test-mixin');
 
 module.exports = TestTask.extend(TestMixin);
