+++ lib/utils/electron-test-mixin.js
